+++ tests/LocalPubSubAdapterTest.php
@@ -48,4 +48,33 @@ class LocalPubSubAdapterTest extends TestCase
 
         $adapter->publish('test_channel', 'This is a message sent to handler1 & handler2');
     }
+
+    public function testPublishBatch()
+    {
+        $adapter = new LocalPubSubAdapter();
+
+        $handler1 = Mockery::mock(\stdClass::class);
+        $handler1->shouldReceive('handle')
+            ->with('This is a message sent to handler1 & handler2')
+            ->once();
+        $handler1->shouldReceive('handle')
+            ->with('This is another message!')
+            ->once();
+        $adapter->subscribe('test_channel', [$handler1, 'handle']);
+
+        $handler2 = Mockery::mock(\stdClass::class);
+        $handler2->shouldReceive('handle')
+            ->with('This is a message sent to handler1 & handler2')
+            ->once();
+        $handler2->shouldReceive('handle')
+            ->with('This is another message!')
+            ->once();
+        $adapter->subscribe('test_channel', [$handler2, 'handle']);
+
+        $messages = [
+            'This is a message sent to handler1 & handler2',
+            'This is another message!',
+        ];
+        $adapter->publishBatch('test_channel', $messages);
+    }
 }