+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## 2.0.0 - 2017-05-16
 
+* Implement new publishBatch method on PubSubAdapterInterface
 * Change message serialization to only json encode and decode
 
 ## 1.0.0 - 2016-09-02